+++ inst/sql/sql_server/CountOutcomes.sql
@@ -21,28 +21,30 @@ limitations under the License.
 {DEFAULT @study_start_date = ''}
 {DEFAULT @study_end_date = ''}
 
-IF OBJECT_ID('tempdb..#counts', 'U') IS NOT NULL
-	DROP TABLE #counts;
+DROP TABLE IF EXISTS #counts;
 	
 CREATE TABLE #counts (
 	outcome_id INT,
 	description VARCHAR(255),
 	outcome_subjects INT,
 	outcome_events INT,
-	outcome_obs_periods INT);
+	outcome_obs_periods INT,
+	observed_days BIGINT);
 	
-INSERT INTO #counts (outcome_id, description, outcome_subjects, outcome_events, outcome_obs_periods)
+INSERT INTO #counts (outcome_id, description, outcome_subjects, outcome_events, outcome_obs_periods, observed_days)
 SELECT outcome_ids.outcome_id,
 	'Outcomes',
 	CASE WHEN outcome_subjects IS NULL THEN 0 ELSE outcome_subjects END AS outcome_subjects,
 	CASE WHEN outcome_events IS NULL THEN 0 ELSE outcome_events END AS outcome_events,
-	CASE WHEN outcome_obs_periods IS NULL THEN 0 ELSE outcome_obs_periods END AS outcome_obs_periods
+	CASE WHEN outcome_obs_periods IS NULL THEN 0 ELSE outcome_obs_periods END AS outcome_obs_periods,
+	CASE WHEN observed_days IS NULL THEN 0 ELSE observed_days END AS observed_days
 FROM #outcome_ids outcome_ids
 LEFT JOIN (
 	SELECT outcome_id, 
 		COUNT(DISTINCT person_id) AS outcome_subjects,
 		COUNT(*) AS outcome_events,
-		COUNT(DISTINCT observation_period_id) AS outcome_obs_periods
+		COUNT(DISTINCT observation_period_id) AS outcome_obs_periods,
+		SUM(observed_days) AS observed_days		
 	FROM #outcomes
 	GROUP BY outcome_id
 	) counts
@@ -54,13 +56,15 @@ SELECT outcome_ids.outcome_id,
 	'Outcomes in study period',
 	CASE WHEN outcome_subjects IS NULL THEN 0 ELSE outcome_subjects END AS outcome_subjects,
 	CASE WHEN outcome_events IS NULL THEN 0 ELSE outcome_events END AS outcome_events,
-	CASE WHEN outcome_obs_periods IS NULL THEN 0 ELSE outcome_obs_periods END AS outcome_obs_periods
+	CASE WHEN outcome_obs_periods IS NULL THEN 0 ELSE outcome_obs_periods END AS outcome_obs_periods,
+	CASE WHEN observed_days IS NULL THEN 0 ELSE observed_days END AS observed_days
 FROM #outcome_ids outcome_ids
 LEFT JOIN (
 	SELECT outcome_id, 
 		COUNT(DISTINCT person_id) AS outcome_subjects,
 		COUNT(*) AS outcome_events,
-		COUNT(DISTINCT observation_period_id) AS outcome_obs_periods
+		COUNT(DISTINCT observation_period_id) AS outcome_obs_periods,
+		SUM(observed_days) AS observed_days		
 	FROM #outcomes_in_period
 	GROUP BY outcome_id
 	) counts
@@ -73,13 +77,15 @@ SELECT outcome_ids.outcome_id,
 	'Outcomes in nesting cohort',
 	CASE WHEN outcome_subjects IS NULL THEN 0 ELSE outcome_subjects END AS outcome_subjects,
 	CASE WHEN outcome_events IS NULL THEN 0 ELSE outcome_events END AS outcome_events,
-	CASE WHEN outcome_obs_periods IS NULL THEN 0 ELSE outcome_obs_periods END AS outcome_obs_periods
+	CASE WHEN outcome_obs_periods IS NULL THEN 0 ELSE outcome_obs_periods END AS outcome_obs_periods,
+	CASE WHEN observed_days IS NULL THEN 0 ELSE observed_days END AS observed_days
 FROM #outcome_ids outcome_ids
 LEFT JOIN (
 	SELECT outcome_id, 
 		COUNT(DISTINCT person_id) AS outcome_subjects,
 		COUNT(*) AS outcome_events,
-		COUNT(DISTINCT observation_period_id) AS outcome_obs_periods
+		COUNT(DISTINCT observation_period_id) AS outcome_obs_periods,
+		SUM(observed_days) AS observed_days		
 	FROM #outcomes_in_nesting
 	GROUP BY outcome_id
 	) counts